BPM_Flash
This sketch is intended to blink a color to a fixed BPM and fade out. 

The sketch takes two arrays of LEDs and the sequence is as follows: 
 
Blue flashes and fades out in one array, then Green in the other, then White on both. 

There will be a momentary push button to cycle between several different BPMs and eventually one to trigger the current loop from the beginning.